- Added virtual memory support to emulator
git-svn-id: svn://localhost/mosa@1454 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/EmulatedKernel/Native.cs
@@ -19,7 +19,7 @@ namespace Mosa.Platforms.x86
 		/// Sets the cr0.
 		/// </summary>
 		/// <param name="state">The state.</param>
-		public static void SetCR0(byte state)
+		public static void SetCR0(uint state)
 		{
 			EmulatedKernel.MemoryDispatch.CR0 = state;
 		}
